Move spider example and use crawler thread runner
