+++ sources/scrapy_scraper/spider.py
@@ -1,7 +1,6 @@
-from typing import Any, Dict, Generator, List, Optional
+from typing import Any, Dict, List, Optional
 
 import scrapy  # type: ignore
-from scrapy.responsetypes import Response  # type: ignore
 
 from .types import BaseQueue
 
@@ -31,47 +30,3 @@ class DLTSpiderBase(scrapy.Spider):
 
     def done(self) -> None:
         self.send_data({"done": True})
-
-
-"""
-Below is a slightly modified example of Scrapy spider
-which you find in the official documentation
-https://docs.scrapy.org/en/latest/intro/tutorial.html#extracting-data-in-our-spider.
-
-To send the data from the spider to pipeline the spider needs a queue instance,
-once scrapy closes the session we also send a termination flag so
-DLT pipeline can stop waiting for more data.
-"""
-
-
-class QuotesSpider(DLTSpiderBase):
-    def parse(
-        self, response: Response
-    ) -> Optional[Generator[scrapy.Request, None, None]]:
-        for quote in response.css("div.quote"):
-            data = {
-                "headers": {
-                    "status": response.status,
-                    **dict(response.headers.to_unicode_dict()),
-                },
-                "quote": {
-                    "text": quote.css("span.text::text").get(),
-                    "author": quote.css("small.author::text").get(),
-                    "tags": quote.css("div.tags a.tag::text").getall(),
-                },
-            }
-
-            # Once we have our data we need to send
-            # it over the queue to pipeline
-            self.send_data(data)
-
-        # Find the next page and create next request otherwise
-        # send termination info.
-        next_page = response.css("li.next a::attr(href)").get()
-        if next_page is not None:
-            next_page = response.urljoin(next_page)
-            yield scrapy.Request(next_page, callback=self.parse)
-        else:
-            # Important: once we are done we need to send signal
-            # to pipeline so it can finalize and shutdown
-            self.done()
